remove uses of 'concat' with a single argument
For example replaced
  [concat $file_name]
with
  $file_name

I'm not sure what the intention of this code was, it seems concat was
consistently used around filenames. Possibly this was a leftover from some
refactoring?

Note that [concat $a] is _not_ the same as $a. So this patch does potentially
change the behavior. See also next commit for more info on 'concat'.