use append instead of concat for strings
'concat' takes multiple arguments, interprets them all as lists and returns a
new list that is obtained by concatenating all elements from the given lists.

'append' takes a name of a variable and zero or more strings. All those strings
are appended to the given variable.

So 'concat' (and 'lappend') is for list manipulation and 'append' is for string
manipulation.

See here for more info:
  http://www.wellho.net/mouth/1402_Tcl-append-v-lappend-v-concat.html