various 'variable' cleanups
Mainly 3 different changes:

1) Some procs mention a variable that's not (or no longer) used inside that
proc. I removed those.

2) Some procs mention a variable that's only used in that proc (not in any
other proc) and whose value is always re-initialized in that proc (so the value
is not reused from an earlier invocation). In other words these can be local
variables instead of global (namespace scoped) variables. An example of this
change is the 'stop_message' variable.

3) Combine 'variable' and 'set' into one where possible. So replace
  variable foo
  set foo "bar"
with
  variable foo "bar"

To make these 3 changes easier to find, I moved the 'variable' declarartion as
close as possible to their point of use instead listing all variable at the top
of the proc.