bugfix in update_frametime ??
The script is hardcoded to sample_accurate mode. And in that case
update_frametime isn't used (so the bug doesn't trigger).

Is it useful to have both modes? If so, should we allow to user to choose the
mode? If not, can we remove the update_frametime and sample_accurate stuff?