simplify watchpoint tracking
The watchpoint-ID's were stored in a dictionary. The keys and values of this
dictionary respectively represent the purpose of the watchpoint and the actual
watchpoint-ID.

But actually we don't need to know the purpose of each watchpoint. The only
reason why we keep this data-structure is to be able to remove the watchpoints.
And for that a list of ID's is sufficient.

This greatly simplifies the 'vgm_rec_end' proc.